requirements.txt
pydactyl
discord.py
python-dotenv
